requirements.txt
# ArbMind dependencies
python-kraken-sdk>=3.2.0       # Kraken REST + WebSocket SDK
langchain>=0.3.0               # LangChain core
langchain-anthropic>=0.3.0     # Claude AI brain
langchain-core>=0.3.0
anthropic>=0.40.0
requests>=2.31.0               # CoinMarketCap API calls
pandas>=2.0.0                  # Correlation matrix, OHLC analysis
numpy>=1.24.0                  # Numerical ops
python-dotenv>=1.0.0           # Environment config
rich>=13.0.0                   # Beautiful terminal output
aiohttp>=3.9.0                 # Async HTTP for CMC
pytz>=2024.1                   # Timezone handling
